Add notebook

+++ metavirommodel/_models.py
@@ -265,7 +265,7 @@ class Metaviromodel(pints.ForwardModel):
 
         """
         # Create timeline vector
-        times = np.arange(start_time, end_time+0.5, 1, dtype=np.int8)
+        times = np.arange(start_time, end_time+0.5, 1, dtype=np.int64)
         interval = end_time - start_time + 1
 
         # Split compartments into their types
@@ -281,7 +281,7 @@ class Metaviromodel(pints.ForwardModel):
         infect_times_history = []
         recov_infect_times_history = []
 
-        solution = np.empty((interval, 3), dtype=np.int8)
+        solution = np.empty((interval, 3), dtype=np.int64)
         S_history = []
         I_history = []
         R_history = []
